PSP init

+++ Source/Core/Src/UnCorSc.cpp
@@ -277,7 +277,7 @@ void UObject::execArrayElement( FFrame& Stack, BYTE*& Result )
 	{
 		// Display out-of-bounds warning and continue on with index clamped to valid range.
 		Stack.ScriptWarn( 0, "Accessed array out of bounds (%i/%i)", *(INT*)Addr, GProperty->ArrayDim );
-		*(INT*)Addr = Clamp( *(INT*)Addr, 0, GProperty->ArrayDim - 1 );
+		*(INT*)Addr = Clamp( *(INT*)Addr, (INT)0, GProperty->ArrayDim - 1 );
 	}
 
 	// Add scaled offset to base pointer.
@@ -2267,8 +2267,8 @@ void UObject::execMid( FFrame& Stack, BYTE*& Result )
 	P_FINISH;
 
 	if( I < 0 ) C += I;
-	I = Clamp( I, 0, (int)appStrlen(A) );
-	C = Clamp( C, 0, (int)appStrlen(A)-I );
+	I = Clamp( I, (INT)0, (int)appStrlen(A) );
+	C = Clamp( C, (INT)0, (int)appStrlen(A)-I );
 	appStrncpy( (char*)Result, A + I, C+1 );
 
 	unguardexecSlow;
@@ -2283,7 +2283,7 @@ void UObject::execLeft( FFrame& Stack, BYTE*& Result )
 	P_GET_INT(N);
 	P_FINISH;
 
-	N = Clamp( N, 0, appStrlen(A) );
+	N = Clamp( N, (INT)0, appStrlen(A) );
 	appStrncpy( (char*)Result, A, N+1 );
 
 	unguardexecSlow;
@@ -2298,7 +2298,7 @@ void UObject::execRight( FFrame& Stack, BYTE*& Result )
 	P_GET_INT(N);
 	P_FINISH;
 
-	N = Clamp( appStrlen(A) - N, 0, appStrlen(A) );
+	N = Clamp( appStrlen(A) - N, (INT)0, appStrlen(A) );
 	appStrcpy( (char*)Result, A + N );
 
 	unguardexecSlow;
